dos.library: Add some comment to prevent confusion about 64-bit support
As noted earlier in the file, 64-bit AROS uses MEMF_31 to make sure
that segments allocated for AOS HUNKs are in the lower address
ranges. This patch adds comments to that effect in some additional
locations, for future maintainers.


git-svn-id: https://svn.aros.org/svn/aros/trunk@37187 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/dos/internalloadseg_aos.c
@@ -292,7 +292,10 @@ BPTR InternalLoadSeg_AOS(BPTR fh,
             //D(bug("\t\t0x%06lx\n", offset));
             addr = (ULONG *)(GETHUNKPTR(curhunk) + offset);
 
-            *addr = AROS_BE2LONG(*addr) + (IPTR)GETHUNKPTR(count);
+            /* See the above MEMF_31 explanation for why this
+             * works on AROS 64-bit.
+             */
+            *addr = (ULONG)(AROS_BE2LONG(*addr) + (IPTR)GETHUNKPTR(count));
 
             --i;
           }
@@ -345,7 +348,10 @@ BPTR InternalLoadSeg_AOS(BPTR fh,
               //D(bug("\t\t0x%06lx\n", offset));
               addr = (ULONG *)(GETHUNKPTR(curhunk) + offset);
 
-              *addr = AROS_BE2LONG(*addr) + (IPTR)GETHUNKPTR(count);
+              /* See the above MEMF_31 explanation for why this
+               * works on AROS 64-bit.
+               */
+              *addr = (ULONG)(AROS_BE2LONG(*addr) + (IPTR)GETHUNKPTR(count));
 
               --i;
             } /* while (i > 0)*/
